Sample Project for Docker implementation
This is sample project in which an environment variable as a string will be provided as input and it will be converted to reverse ordered string  and will be saved to text file. 
The DockerFile added to write command to build and publish the project. The docker-compose.yml file added to set service, volumes etc.